Added the new networking framing model, so that information can be sent properly. Please see the testing.py file to know how the client will send and respond to messages from the server
