+++ .gitignore
@@ -161,3 +161,6 @@ cython_debug/
 #  option (not recommended) you can uncomment the following to ignore the entire idea folder.
 .idea/
 .vscode
+*.m4a
+*.txt
+.DS_Store
